Added a check to Interaction.cs so that the marking menu would still work on vive.
Added the depthPeelingRenderer (currently in progress) in case we want to
go in that direction. Work stalled because Eric found an easier
workaround.

Added a check to VrGame.cs so that the program would not crash if OVR was
not running (for instance on my debug machine, the Vive.)

Changed the visibility of several of the fields in VrRenderer.cs to
protected (from private) in order to access them in
DepthPeelingRenderer.cs, which extends VrRenderer.